+++ src/api/catalog/categories/all.ts
