+++ src/api/catalog/categories/delete.ts
