+++ src/api/catalog/categories/single.ts
