--- src/api/categories/common.ts
@@ -1,7 +1,0 @@
-export interface CategoryResponse {
-	id: number;
-	name: string;
-	description: string;
-}
-
-export const CATEGORY_BASE_PATH = '/categories';
