--- src/api/customizations/materials/all.ts
@@ -1,3 +1,0 @@
-import { MATERIALS_BASE_PATH } from '@/api/customizations/common';
-
-export const url = () => `${MATERIALS_BASE_PATH}`;
