+++ src/api/printing/common.ts
