+++ src/api/printing/customizations/create.ts
@@ -1,5 +1,5 @@
 import * as headers from '@/types/headers';
-import { CUSTOMIZATIONS_BASE_PATH } from '@/api/customizations/common';
+import { CUSTOMIZATIONS_BASE_PATH } from '@/api/printing/common';
 
 export type Request = {
 	scale: number;
