+++ src/api/printing/materials/all.ts
@@ -1,0 +1,3 @@
+import { MATERIALS_BASE_PATH } from '@/api/printing/common';
+
+export const url = () => `${MATERIALS_BASE_PATH}`;
