+++ src/api/printing/materials/delete.ts
@@ -1,4 +1,4 @@
-import { MATERIALS_BASE_PATH } from '@/api/customizations/common';
+import { MATERIALS_BASE_PATH } from '@/api/printing/common';
 
 export interface Request {
 	id: number;
