+++ src/api/printing/materials/download.ts
@@ -1,4 +1,4 @@
-import { MATERIALS_BASE_PATH } from '@/api/customizations/common';
+import { MATERIALS_BASE_PATH } from '@/api/printing/common';
 import { DownloadResponse } from '@/api/common/files';
 
 export interface Request {
