+++ src/api/printing/materials/replace.ts
@@ -1,4 +1,4 @@
-import { MATERIALS_BASE_PATH } from '@/api/customizations/common';
+import { MATERIALS_BASE_PATH } from '@/api/printing/common';
 import { UploadRequest } from '@/api/common/files';
 
 export interface Request {
