+++ src/api/printing/materials/upload.ts
@@ -1,4 +1,4 @@
-import { MATERIALS_BASE_PATH } from '@/api/customizations/common';
+import { MATERIALS_BASE_PATH } from '@/api/printing/common';
 import { UploadRequest, UploadResponse } from '@/api/common/files';
 
 export interface Request {
