+++ src/app/components/search/categories.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { faChevronDown } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { CategoryResponse } from '@/api/categories/common';
+import { CategoryResponse } from '@/api/catalog/common';
 import { useGetCategories } from '@/hooks/queries/categories';
 import Loader from '@/app/components/state/loading';
 import ErrorPage from '@/app/components/state/error';
